samples: boards: rpi_pico: Added spi_pio sample
Added a sample for the rpi_pico board demonstrating using the SPI over
PIO driver.  The provided rpi_pico.overlay uses pins 12 through 15 for
the SPI interface.

This sample is based on the zephyr/samples/sensor/bme280 sample;  the
only changes are adding the rpi_pico.overlay and updating the prj.cfg.

Signed-off-by: Steve Boylan <stephen.boylan@beechwoods.com>